checking if file exists before removing markers

+++ bundles/org.eclipse.wst.sse.ui/src/org/eclipse/wst/sse/ui/internal/search/BasicSearchQuery.java
@@ -179,8 +179,10 @@ public class BasicSearchQuery implements ISearchQuery {
 
 	public Match[] getMatches() {
 		// get rid of the old markers
-		deleteOccurrencesMarkers();
-		doQuery();
+		if(fFile.exists()) {
+			deleteOccurrencesMarkers();
+			doQuery();
+		}
 		return (Match[]) this.fMatches.toArray(new Match[this.fMatches.size()]);
 	}
